tools/scripts/px4_uml_viewer/requirements.txt
flake8==4.0.1
pyYAML==6.0
